project-int

+++ src/main/java/com/sh/sh/pos/system/model/ShiftReport.java
@@ -40,7 +40,7 @@ public class ShiftReport {
 	
 	private Double netSale;
 	
-	private Double totalOrder;
+	private int totalOrder;
 	
 	@ManyToOne
 	private User cashier ;
